Moved VVHClientSpec test file and removed unnecessary test
