+++ digiroad2-oracle/src/main/scala/fi/liikennevirasto/digiroad2/VVHClient.scala
@@ -799,17 +799,6 @@ class VVHHistoryClient(vvhRestApiEndPoint: String) extends VVHClient(vvhRestApiE
     Future(fetchVVHRoadlinkHistoryByBoundsAndMunicipalities(bounds, municipalities))
   }
 
-  /**
-    * For testing purposes
-    * @param bounds
-    * @param municipalities
-    * @return
-    */
-  def uRLgenerator(bounds: BoundingRectangle,municipalities: Set[Int]) :String ={
-    historyURLGenerator(bounds:BoundingRectangle,municipalities: Set[Int])
-  }
-
-
   private def makeFilter[T](attributeName: String, ids: Set[T]): String = {
     val filter =
       if (ids.isEmpty) {
@@ -867,17 +856,14 @@ class VVHHistoryClient(vvhRestApiEndPoint: String) extends VVHClient(vvhRestApiE
   * PointAssetService.getByBoundingBox and ServicePointImporter.importServicePoints.
   */
   def fetchVVHRoadlinkHistoryByBoundsAndMunicipalities(bounds: BoundingRectangle, municipalities: Set[Int] = Set()): Seq[VVHRoadlink] = {
-   val url= historyURLGenerator(bounds,municipalities)
-      fetchVVHFeatures(url) match {
-        case Left(features) => features.map(extractVVHFeature)
-        case Right(error) => throw new VVHClientException(error.toString)
-      }
-    }
-  def historyURLGenerator(bounds:BoundingRectangle,municipalities: Set[Int]):String ={
     val definitions = historyLayerDefinition(withMunicipalityFilter(municipalities))
-    vvhRestApiEndPoint + roadLinkDataHistoryService + "/FeatureServer/query?" +
+    val url= vvhRestApiEndPoint + roadLinkDataHistoryService + "/FeatureServer/query?" +
     s"layerDefs=$definitions&geometry=" + bounds.leftBottom.x + "," + bounds.leftBottom.y + "," + bounds.rightTop.x + "," + bounds.rightTop.y +
     "&geometryType=esriGeometryEnvelope&spatialRel=esriSpatialRelIntersects&" + queryParameters()
+    fetchVVHFeatures(url) match {
+      case Left(features) => features.map(extractVVHFeature)
+      case Right(error) => throw new VVHClientException(error.toString)
+    }
   }
 }
